mudando para projeto java normal
