+++ .recommenders/caches/identified-project-coordinates.json
@@ -1,1 +1,1 @@
-[[{"location":"C:\\Program Files\\Java\\jre1.8.0_171","type":"JRE","hints":{"EXECUTION_ENVIRONMENT":"J2SE-1.5"}},"jre:jre:1.5.0"]]
+{}
