+++ comp2/.settings/org.jboss.ide.eclipse.as.core.prefs
@@ -1,0 +1,2 @@
+eclipse.preferences.version=1
+org.jboss.ide.eclipse.as.core.singledeployable.deployableList=
