+++ comp2/src/animais/Leao.java
@@ -1,0 +1,13 @@
+package animais;
+
+import animais.abstracts.TerrrestreAbstract;
+import animais.enums.Alimento;
+import animais.enums.Porte;
+import animais.enums.Sexo;
+
+public class Leao extends TerrrestreAbstract {
+	public Leao(Sexo sexo,Porte porte){		
+		super(sexo,porte);
+		this.alimentosConsumiveis.add(Alimento.CARNES);
+	}	
+}
