+++ comp2/src/animais/enums/Alimento.java
@@ -1,0 +1,5 @@
+package animais.enums;
+
+public enum Alimento {	
+	FRUTAS, VERDURAS, CARNES, PEIXES;
+}
