+++ comp2/src/animais/enums/Porte.java
@@ -1,0 +1,5 @@
+package animais.enums;
+
+public enum Porte {	
+	PEQUENO, MEDIO, GRANDE;
+}
