+++ comp2/src/animais/enums/Sexo.java
@@ -1,0 +1,5 @@
+package animais.enums;
+
+public enum Sexo {	
+	MASCULINO, FEMININO;
+}
