+++ comp2/src/animais/interfaces/Anfibio.java
@@ -1,0 +1,4 @@
+package animais.interfaces;
+
+public interface Anfibio extends Terrestre,Aquatico {
+}
