+++ comp2/src/animais/interfaces/Aquatico.java
@@ -1,0 +1,5 @@
+package animais.interfaces;
+
+public interface Aquatico extends Animal {
+	public void nadar();	
+}
