+++ comp2/src/animais/interfaces/Terrestre.java
@@ -1,0 +1,6 @@
+package animais.interfaces;
+
+public interface Terrestre extends Animal {
+	public void andar();
+	
+}
