+++ comp2/src/instalacao/enums/Localizacao.java
@@ -1,0 +1,5 @@
+package instalacao.enums;
+
+public enum Localizacao {
+	NORTE, SUL, LESTE, OESTE,NORDESTE,NOROESTE,SULDESTE,SULDOESTE, CENTROESTE;
+}
