+++ comp2/src/instalacao/tipos/ElementoInstalacao.java
@@ -1,0 +1,5 @@
+package instalacao.tipos;
+
+public class ElementoInstalacao {
+
+}
