--- comp2/src/main/java/animais/enums/Alimento.java
@@ -1,5 +1,0 @@
-package animais.enums;
-
-public enum Alimento {	
-	FRUTAS, VERDURAS, CARNES, PEIXES;
-}
