--- comp2/src/main/java/animais/enums/Porte.java
@@ -1,5 +1,0 @@
-package animais.enums;
-
-public enum Porte {	
-	PEQUENO, MEDIO, GRANDE;
-}
