--- comp2/src/main/java/animais/enums/Sexo.java
@@ -1,5 +1,0 @@
-package animais.enums;
-
-public enum Sexo {	
-	MASCULINO, FEMININO;
-}
