--- comp2/src/main/java/animais/tipos/Anfibio.java
@@ -1,4 +1,0 @@
-package animais.tipos;
-
-public interface Anfibio extends Animal {
-}
