--- comp2/src/main/java/animais/tipos/Animal.java
@@ -1,14 +1,0 @@
-package animais.tipos;
-
-import animais.enums.Alimento;
-import animais.enums.Porte;
-import animais.enums.Sexo;
-
-public interface Animal {	
-	public boolean alimentar(Alimento alimento);
-	public void locomover();
-	public Sexo getSexo();
-	public void setSexo(Sexo sexo);
-	public Porte getPorte();
-	public void setPorte(Porte porte);
-}
