--- comp2/src/main/java/animais/tipos/Aquatico.java
@@ -1,4 +1,0 @@
-package animais.tipos;
-
-public interface Aquatico extends Animal {
-}
