--- comp2/src/main/java/animais/tipos/Terrestre.java
@@ -1,4 +1,0 @@
-package animais.tipos;
-
-public interface Terrestre extends Animal {
-}
