--- comp2/src/main/java/instalacao/enums/Localizacao.java
@@ -1,5 +1,0 @@
-package instalacao.enums;
-
-public enum Localizacao {
-	NORTE, SUL, LESTE, OESTE,NORDESTE,NOROESTE,SULDESTE,SULDOESTE, CENTROESTE;
-}
