--- comp2/src/main/java/instalacao/tipos/ElementoInstalacao.java
@@ -1,5 +1,0 @@
-package instalacao.tipos;
-
-public class ElementoInstalacao {
-
-}
